requirements.txt
# Project dependencies
# Add specific versions as needed, e.g. package==1.0.0

strands-agents>=0.1.0
strands-agents-tools>=0.1.0
requests>=2.28.0
boto3>=1.28.0

# OpenTelemetry dependencies
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-logging>=0.40b0
opentelemetry-instrumentation-requests>=0.40b0
